+++ safespaces/devices/rift.lua
@@ -1,0 +1,24 @@
+return {
+	display = '^Rift/WMHD',
+	oversample_w = 1.4,
+	oversample_h = 1.4,
+	distortion_model = "basic",
+	prelaunch = true
+-- metadata overrides
+-- width = .. (0.5width*oversample_w)
+-- height = ..
+-- center = ..
+-- horizontal = ..
+-- vertical = ..,
+-- left_fov = ..
+-- right_fov = ..
+-- left_ar = ..
+-- right_ar = ..
+-- hsep = ..
+-- vpos = ..
+-- lens_distance = ..
+-- eye_display = ..
+-- ipd = ..
+-- distortion = {v1, v2, v3, v4}
+-- abberation = {v1, v2, v3}
+};