+++ CHANGELOG.md
@@ -1,1 +1,6 @@
+0.2 (ongoing)
+
+* Added 'prelaunch' option to device profiles
+* Added Rift device profile
+
 0.1 : 28/3/2018 - First Release